backend/requirements.txt
Django>=4.2.0,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
Pillow>=9.0.0
python-decouple>=3.6
django-filter>=23.0
reportlab>=4.0.0
fonttools>=4.0.0